Move the parts of the build that are not a kind of page
build.py is one function per kind of page, and it had grown four hundred
lines that are not a kind of page. Reading it to change how the hub
renders meant loading the sitemap writer, the Atom feeds, the emitter
and the whole of --check first. buildlib/ was already the answer to this
question - templating, minifying, i18n, imports each have a file - and
these three groups had simply never been asked to leave.

  buildlib/catalogue.py  sitemap.xml, llms.txt and the Atom feeds. The
    same job with three audiences: each is built from the same two lists
    and each holds back the same languages, because a locale still being
    translated carries no hreflang tag and no switcher entry and a line
    in any of these would be the last remaining way to send a crawler at
    a page that is half in English. Getting that right in three places
    and wrong in a fourth is the mistake this is meant to make hard, and
    one file is most of how. None of the three is a page: no frame, no
    stylesheet, no worker, no locale to be rendered in.

  buildlib/emit.py  write, LINK and Emitter - the one place a string
    becomes a file. `write` earns its own home by what it refuses: every
    text file here is LF, and write_text() writes CRLF on Windows, which
    would make the deployed site differ from a local build on every line
    of every file on alternate machines. LINK goes with the emitter
    rather than with check_links because gathering a page's links is
    something the emitter does AS it writes; check_links borrows the
    pattern for the path where it re-reads.

  buildlib/deployed.py  --check. It runs after a build rather than
    during one, nothing else calls it, and it was eighty lines that only
    execute when somebody is verifying a deploy.

The tests follow the code: buildlib/X.py has always had
tests/python/test_X.py beside it, so Write, EmitterSetup and
EmitterOutput are now test_emit.py and BlobId is test_deployed.py. Same
assertions, same names, moved.

Nothing here changes what is built, and that is checked rather than
claimed: a full build before this commit and one after produce 11,543
files with zero differing bytes.
